fixing negative event numbers. Details: https://its.cern.ch/jira/browse/ATLASVPONE-287. (VP1Gui-00-08-11)
    * fixing negative event numbers.
	  Details: https://its.cern.ch/jira/browse/ATLASVPONE-287
	* new tag: VP1Gui-00-08-11


Former-commit-id: 28e17a5c888d71f2093b5fc4f52a4244a564fa9a

+++ graphics/VP1/VP1Gui/VP1Gui/VP1EvtsOnServerInfo.h
@@ -35,7 +35,7 @@ public:
   QStringList print() const;
 
   //Access the results:
-  int numberOfEvents() const;
+  unsigned long long numberOfEvents() const;
   VP1EventFile newestEvent() const;
   QList<VP1EventFile> events(int timecut = -1, bool requireNewestRunNumber = false ) const;
   //Returns events in order of increasing time.
